feat: scheduler + parameters for 1st experiment
