+++ dataloader/__init__.py
@@ -1,0 +1,3 @@
+from .fmnist import *
+from .mnist import *
+from .permutedmnist import *
