+++ dataloader/permutedmnist.py
@@ -4,72 +4,6 @@ import torch
 import random
 
 
-def mnist(path, batch_size):
-    """ Load MNIST dataset
-
-    Args:
-        path (str): Path to save/load dataset
-        batch_size (int): Batch size for training
-
-    Returns:
-        torch.utils.data.DataLoader: MNIST training dataset
-        torch.utils.data.DataLoader: MNIST testing dataset
-    """
-    ### NORMALIZATION ###
-    mean, std = 0.1307, 0.3081  # As in EWC paper
-    normalisation = transforms.Compose([
-        transforms.ToTensor(),
-        transforms.Normalize(mean, std)
-    ])
-
-    ### DOWNLOAD DATASETS ###
-    mnist_train = datasets.MNIST(
-        root=path, download=True, transform=normalisation, target_transform=None, train=True)
-    mnist_test = datasets.MNIST(
-        root=path, download=True, transform=normalisation, target_transform=None, train=False)
-
-    ### DATA LOADER ###
-    mnist_train = DataLoader(
-        mnist_train, batch_size=batch_size, shuffle=True)
-    mnist_test = DataLoader(
-        mnist_test, batch_size=batch_size, shuffle=False)
-
-    return mnist_train, mnist_test
-
-
-def fashion_mnist(path, batch_size):
-    """ Load Fashion-MNIST dataset
-
-    Args:
-        path (str): Path to save/load dataset
-        batch_size (int): Batch size for training
-
-    Returns:
-        torch.utils.data.DataLoader: Fashion-MNIST training dataset
-        torch.utils.data.DataLoader: Fashion-MNIST testing dataset
-    """
-    ### NORMALIZATION ###
-    mean, std = 0.1307, 0.3081  # As in EWC paper
-    normalisation = transforms.Compose([
-        transforms.ToTensor(),
-        transforms.Normalize(mean, std)
-    ])
-
-    ### DOWNLOAD DATASETS ###
-    fashion_mnist_train = datasets.FashionMNIST(
-        root=path, download=True, transform=normalisation, target_transform=None, train=True)
-    fashion_mnist_test = datasets.FashionMNIST(
-        root=path, download=True, transform=normalisation, target_transform=None, train=False)
-
-    ### DATA LOADER ###
-    fashion_mnist_train = DataLoader(
-        fashion_mnist_train, batch_size=batch_size, shuffle=True)
-    fashion_mnist_test = DataLoader(
-        fashion_mnist_test, batch_size=batch_size, shuffle=False)
-
-    return fashion_mnist_train, fashion_mnist_test
-
-
 class PermutedMNIST(datasets.MNIST):
     """ Permuted MNIST dataset
