+++ utils/visual.py
@@ -17,10 +17,15 @@ def versionning(folder, title, format=".pdf"):
 
 def visualize_sequential(title, t_accuracy, folder):
     plt.figure()
-    plt.plot(range(1, len(t_accuracy)+1), t_accuracy)
+    plt.plot(range(1, len(t_accuracy)+1), t_accuracy, zorder=3)
     plt.xlabel('Epoch')
     plt.ylabel('Accuracy')
-    plt.legend(['MNIST', 'Fashion-MNIST'])
+    plt.legend([f"Task {i+1}" for i in range(len(t_accuracy))])
+    n_epochs_task = len(t_accuracy) // len(t_accuracy[0])
+    # Vertical lines to separate tasks
+    for i in range(1, len(t_accuracy[0])):
+        plt.axvline(x=i*n_epochs_task, color='k',
+                    linestyle='--', linewidth=0.5)
     os.makedirs(folder, exist_ok=True)
     versionned = versionning(folder, title)
     plt.savefig(versionned, bbox_inches='tight')
